Fixed Minimum Speed, Generation Error, Lane Switching
Players can now only switch to a third lane if the road they will be on has a third lane.

Minimum speed is not enforced for a certain amount of time at the beginning of the game.

Generation works, but I need to get over the work I did in my last branch.